# Admin/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specifiction - APPLYCAM",
    "description": "OpenApi documentation for applycam project",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "jotsamikael",
      "url": "https://dodibo.com/",
      "email": "jotsamikael@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://someurl.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV."
    },
    {
      "url": "https://something:8088/api/v1",
      "description": "Prod ENV."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/payment/create-payment": {
      "post": {
        "tags": [
          "payment"
        ],
        "operationId": "createPayment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/campus": {
      "post": {
        "tags": [
          "campus"
        ],
        "operationId": "createCampus",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CreateCampusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/payment/update-payment/{id}": {
      "patch": {
        "tags": [
          "payment"
        ],
        "operationId": "updatePayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/campus/update-campus": {
      "patch": {
        "tags": [
          "campus"
        ],
        "operationId": "updateCampus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/campus/delete-campus/{name}": {
      "patch": {
        "tags": [
          "campus"
        ],
        "operationId": "deleteCampus",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/subjects/findByName/{name}": {
      "get": {
        "tags": [
          "subject"
        ],
        "operationId": "findByName",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/specialities/findByName/{name}": {
      "get": {
        "tags": [
          "speciality"
        ],
        "operationId": "findByName_1",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialityResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/session/findBy-examdate/{examDate}": {
      "get": {
        "tags": [
          "session"
        ],
        "operationId": "findByName_2",
        "parameters": [
          {
            "name": "examDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/payment/get-all": {
      "get": {
        "tags": [
          "payment"
        ],
        "operationId": "getAllPayments",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePaymentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/course-management/findByName/{name}": {
      "get": {
        "tags": [
          "course"
        ],
        "operationId": "findByName_3",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/campus/get-campus-by-training-town/{town}": {
      "get": {
        "tags": [
          "campus"
        ],
        "operationId": "findCampusByTown",
        "parameters": [
          {
            "name": "town",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCampusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/campus/get-campus-by-training-center/{agreementNumber}": {
      "get": {
        "tags": [
          "campus"
        ],
        "operationId": "findCampusByTrainingCenter",
        "parameters": [
          {
            "name": "agreementNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampusResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/campus/find-campus/{name}": {
      "get": {
        "tags": [
          "campus"
        ],
        "operationId": "findCampusByName",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/assignment/findBy-division/{division}": {
      "get": {
        "tags": [
          "exam-center-controller"
        ],
        "operationId": "findByName_4",
        "parameters": [
          {
            "name": "division",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseExamCenterResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/payment/delete/{id}": {
      "delete": {
        "tags": [
          "payment"
        ],
        "operationId": "deletePayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatePaymenRequest": {
        "required": [
          "amount",
          "paymentMethod",
          "secretCode"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "paymentMethod": {
            "type": "string"
          },
          "secretCode": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateCampusRequest": {
        "required": [
          "capacity",
          "name",
          "quarter",
          "town",
          "trainingCenterAgr"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "quarter": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "trainingCenterAgr": {
            "type": "string"
          },
          "xcoor": {
            "type": "number",
            "format": "double"
          },
          "ycoor": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "UpdateCampusRequest": {
        "required": [
          "capacity",
          "name",
          "oldName",
          "quarter",
          "town"
        ],
        "type": "object",
        "properties": {
          "oldName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "quarter": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "xcoor": {
            "type": "number",
            "format": "double"
          },
          "ycoor": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "Campus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "quarter": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "trainingCenter": {
            "$ref": "#/components/schemas/TrainingCenter"
          },
          "xcoor": {
            "type": "number",
            "format": "double"
          },
          "ycoor": {
            "type": "number",
            "format": "double"
          },
          "actived": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "HasSchooled": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "startYear": {
            "type": "string",
            "format": "date"
          },
          "endYear": {
            "type": "string",
            "format": "date"
          },
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "trainingCenter": {
            "$ref": "#/components/schemas/TrainingCenter"
          },
          "actived": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "OffersSpeciality": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "agreement": {
            "type": "string"
          },
          "speciality": {
            "$ref": "#/components/schemas/Speciality"
          },
          "trainingCenter": {
            "$ref": "#/components/schemas/TrainingCenter"
          },
          "actived": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "TrainingCenter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "acronym": {
            "type": "string"
          },
          "agreementNumber": {
            "type": "string"
          },
          "agreementFileUrl": {
            "type": "string"
          },
          "agreementStatus": {
            "type": "string"
          },
          "startDateOfAgreement": {
            "type": "string",
            "format": "date"
          },
          "endDateOfAgreement": {
            "type": "string",
            "format": "date"
          },
          "isCenterPresentCandidateForCqp": {
            "type": "boolean"
          },
          "isCenterPresentCandidateForDqp": {
            "type": "boolean"
          },
          "division": {
            "type": "string"
          },
          "fullAddress": {
            "type": "string"
          },
          "centerType": {
            "type": "string"
          },
          "centerPhone": {
            "type": "string"
          },
          "centerEmail": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "centerAge": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/TrainingCenterStatusHistory"
          },
          "offersSpecialityList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OffersSpeciality"
            }
          },
          "campusList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Campus"
            }
          },
          "hasSchooledList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HasSchooled"
            }
          },
          "actived": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "TrainingCenterStatusHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64"
          },
          "lastModifiedBy": {
            "type": "integer",
            "format": "int64"
          },
          "trainingCenter": {
            "$ref": "#/components/schemas/TrainingCenter"
          },
          "status": {
            "type": "string",
            "enum": [
              "READYTOPAY",
              "DRAFT",
              "PAID",
              "VALIDATED",
              "INCOMPLETED",
              "REJECTED",
              "PENDING"
            ]
          },
          "comment": {
            "type": "string"
          },
          "actived": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "SubjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "SpecialityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "examType": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "examType": {
            "type": "string"
          },
          "examDate": {
            "type": "string",
            "format": "date"
          },
          "sessionYear": {
            "type": "string"
          }
        }
      },
...
      "PageResponsePaymentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "paymentMethod": {
            "type": "string"
          }
        }
      },
...
      "CourseResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "CampusResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "quarter": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "trainingCenterCampus": {
            "$ref": "#/components/schemas/TrainingCenter"
          },
          "xcoor": {
            "type": "number",
            "format": "double"
          },
          "ycoor": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PageResponseCampusResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampusResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ExamCenterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "division": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseExamCenterResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamCenterResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}